Pull control panel implementation from GLViewerControl to parent class
