+++ GUI/Controls/GLViewerControl.cs
@@ -13,13 +13,12 @@ using WinFormsMouseEventArgs = System.Windows.Forms.MouseEventArgs;
 
 namespace GUI.Controls
 {
-    partial class GLViewerControl : UserControl
+    partial class GLViewerControl : ControlPanelView
     {
+        protected override Panel ControlsPanel => controlsPanel;
         static readonly TimeSpan FpsUpdateTimeSpan = TimeSpan.FromSeconds(0.1);
 
         public GLControl GLControl { get; }
-
-        private int currentControlsHeight = 10;
 
         public struct RenderEventArgs
         {
@@ -55,7 +54,6 @@ namespace GUI.Controls
         public GLViewerControl(VrfGuiContext guiContext)
         {
             InitializeComponent();
-            Dock = DockStyle.Fill;
 
             Camera = new Camera();
 
@@ -178,88 +176,6 @@ namespace GUI.Controls
             form.FormClosed -= OnFullScreenFormClosed;
 
             FullScreenForm = null;
-        }
-
-        public void AddControl(Control control)
-        {
-            controlsPanel.Controls.Add(control);
-            SetControlLocation(control);
-        }
-
-        public CheckBox AddCheckBox(string name, bool defaultChecked, Action<bool> changeCallback)
-        {
-            var checkbox = new GLViewerCheckboxControl(name, defaultChecked);
-            checkbox.CheckBox.CheckedChanged += (_, __) =>
-            {
-                changeCallback(checkbox.CheckBox.Checked);
-            };
-
-            controlsPanel.Controls.Add(checkbox);
-
-            SetControlLocation(checkbox);
-
-            return checkbox.CheckBox;
-        }
-
-        public ComboBox AddSelection(string name, Action<string, int> changeCallback)
-        {
-            var selectionControl = new GLViewerSelectionControl(name);
-
-            controlsPanel.Controls.Add(selectionControl);
-
-            SetControlLocation(selectionControl);
-
-            selectionControl.ComboBox.SelectedIndexChanged += (_, __) =>
-            {
-                selectionControl.Refresh();
-                changeCallback(selectionControl.ComboBox.SelectedItem as string, selectionControl.ComboBox.SelectedIndex);
-            };
-
-            return selectionControl.ComboBox;
-        }
-
-        public CheckedListBox AddMultiSelection(string name, Action<CheckedListBox> initializeCallback, Action<IEnumerable<string>> changeCallback)
-        {
-            var selectionControl = new GLViewerMultiSelectionControl(name);
-
-            initializeCallback?.Invoke(selectionControl.CheckedListBox);
-
-            controlsPanel.Controls.Add(selectionControl);
-
-            SetControlLocation(selectionControl);
-
-            selectionControl.CheckedListBox.ItemCheck += (_, __) =>
-            {
-                // ItemCheck is called before CheckedItems is updated
-                BeginInvoke((MethodInvoker)(() =>
-                {
-                    selectionControl.Refresh();
-                    changeCallback(selectionControl.CheckedListBox.CheckedItems.OfType<string>());
-                }));
-            };
-
-            return selectionControl.CheckedListBox;
-        }
-
-        public GLViewerTrackBarControl AddTrackBar(Action<int> changeCallback)
-        {
-            var trackBar = new GLViewerTrackBarControl();
-            trackBar.TrackBar.Scroll += (_, __) =>
-            {
-                changeCallback(trackBar.TrackBar.Value);
-            };
-
-            controlsPanel.Controls.Add(trackBar);
-
-            SetControlLocation(trackBar);
-
-            return trackBar;
-        }
-
-        public void SetControlLocation(Control control)
-        {
-            control.Location = new Point(0, currentControlsHeight);
-            currentControlsHeight += control.Height;
         }
 
         private void OnDisposed(object sender, EventArgs e)
